requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
alembic==1.18.4
altair==6.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
asgiref==3.11.0
asyncpg==0.31.0
attrs==25.4.0
Authlib==1.6.9
bcrypt==5.0.0
blinker==1.9.0
cachetools==6.2.6
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
colorama==0.4.6
cryptography==46.0.6
dataclasses-json==0.6.7
distlib==0.4.0
distro==1.9.0
dj-database-url==3.0.1
ecdsa==0.19.1
faiss-cpu==1.13.2
fastapi==0.135.3
filelock==3.20.1
fsspec==2026.2.0
gitdb==4.0.12
GitPython==3.1.46
greenlet==3.3.2
grpcio==1.78.0
gunicorn==23.0.0
h11==0.16.0
h2==4.3.0
hf-xet==1.3.2
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.6.0
hyperframe==6.1.0
idna==3.11
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.3
jose==1.0.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
langchain==1.2.12
langchain-classic==1.0.2
langchain-community==0.4.1
langchain-core==1.2.19
langchain-openai==1.1.11
langchain-qdrant==1.1.0
langchain-text-splitters==1.1.1
langgraph==1.1.1
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.11
langsmith==0.7.16
llvmlite==0.46.0
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
mdurl==0.1.2
more-itertools==10.8.0
mpmath==1.3.0
multidict==6.7.1
mypy_extensions==1.1.0
mysqlclient==2.2.7
narwhals==2.16.0
networkx==3.6.1
numba==0.64.0
numpy==2.4.2
ollama==0.6.1
openai==2.30.0
openai-whisper==20250625
orjson==3.11.7
ormsgpack==1.12.2
packaging==25.0
pandas==2.3.3
passlib==1.7.4
pdfminer.six==20231228
pdfplumber==0.11.4
pillow==12.0.0
platformdirs==4.5.1
pluggy==1.6.0
portalocker==3.2.0
propcache==0.4.1
protobuf==6.33.5
psycopg2-binary==2.9.11
pyarrow==23.0.0
pyasn1==0.6.2
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
pydeck==0.9.1
pyee==13.0.1
Pygments==2.19.2
pypdf==6.8.0
pypdfium2==5.6.0
pytest==9.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-jose==3.3.0
python-multipart==0.0.9
pytz==2025.2
pywin32==311
PyYAML==6.0.3
qdrant-client==1.17.0
rank-bm25==0.2.2
redis==7.3.0
referencing==0.37.0
regex==2026.3.32
requests==2.33.1
requests-toolbelt==1.0.0
rich==14.3.3
rpds-py==0.30.0
rsa==4.9.1
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
sentence-transformers==5.3.0
setuptools==82.0.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.48
sqlparse==0.5.4
starlette==1.0.0
streamlit==1.54.0
sympy==1.14.0
tenacity==9.1.3
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.2
toml==0.10.2
torch==2.10.0
torchaudio==2.10.0+xpu
torchvision==0.25.0+xpu
tornado==6.5.4
tqdm==4.67.3
transformers==5.3.0
typer==0.24.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uuid==1.30
uuid_utils==0.14.1
uvicorn==0.42.0
virtualenv==20.36.1
watchdog==6.0.0
Werkzeug==3.1.6
whitenoise==6.11.0
xxhash==3.6.0
yarl==1.23.0
zstandard==0.25.0
